Silence compiler warning
test/generate_lpf.cpp:40:24: error:
      comparison of integers of different signs: 'size_t' (aka
'unsigned long')
      and 'int' [-Werror,-Wsign-compare]
  for (size_t i = 2; i <= max; i++)

+++ test/generate_lpf.cpp
@@ -37,13 +37,13 @@ int main()
   auto lpf = generate_lpf(max);
   auto primes = generate_primes<int>(max);
 
-  for (size_t i = 2; i <= max; i++)
+  for (int i = 2; i <= max; i++)
   {
     int factor = i;
     int sqrt = isqrt(i);
 
     // find smallest prime factor
-    for (size_t j = 1; primes[j] <= sqrt; j++)
+    for (int j = 1; primes[j] <= sqrt; j++)
     {
       if (i % primes[j] == 0)
       {
